perf: avoid re-render of Exchange in ListSymbols
