+++ src/pages/Dashboard/components/ListSymbols/components/ExchangeTable/index.tsx
@@ -1,0 +1,89 @@
+import { useState, useCallback, useEffect } from "react";
+import * as S from "../../styles";
+import useWebSocket from "react-use-websocket";
+import { useExchangeInfo } from "../../../../context/useExchangeInfo";
+
+interface WebSocketMessage {
+  stream: string;
+  data: any;
+  // TODO: increase more details about object WebSocketMessage
+}
+
+const ExchangeTable = () => {
+  const [exchangesInfo, setExchangesInfo] = useState<any[]>([]);
+
+  const { exchanges } = useExchangeInfo();
+
+  const symbolList =
+    exchanges.currentList !== "" &&
+    Object.values(exchanges.lists[exchanges.currentList])
+      ?.map((symbol: any) => symbol.symbol.toLowerCase())
+      .join("@ticker/");
+
+  const wsUrl = `wss://stream.binance.com:9443/stream?streams=${symbolList}`;
+
+  const { lastJsonMessage } = useWebSocket<WebSocketMessage>(wsUrl, {
+    onOpen: () => console.log(`Connected to App WS`),
+    onError: (event) => {
+      console.error(event);
+    },
+    onMessage: () => {
+      if (lastJsonMessage) {
+        setExchangesInfo((prev) => ({
+          ...prev,
+          [lastJsonMessage.stream]: { ...lastJsonMessage.data },
+        }));
+      }
+    },
+    shouldReconnect: () => true,
+    reconnectInterval: 10,
+  });
+
+  const formatNumberToFixed = useCallback(
+    (value: string, decimalPlaces: number = 4) => {
+      return parseFloat(value).toFixed(decimalPlaces);
+    },
+    []
+  );
+
+  const formatPercentage = useCallback((value: number) => {
+    return (value * 10000).toFixed(2);
+  }, []);
+
+  useEffect(() => {
+    setExchangesInfo([]);
+
+    return () => {
+      setExchangesInfo([]);
+    };
+  }, [exchanges.currentList]);
+
+  return (
+    <S.Table>
+      <S.TableHead>
+        <tr>
+          <S.TableHeaderCell>Symbol</S.TableHeaderCell>
+          <S.TableHeaderCell>Last Price</S.TableHeaderCell>
+          <S.TableHeaderCell>Bid Price</S.TableHeaderCell>
+          <S.TableHeaderCell>Ask Price</S.TableHeaderCell>
+          <S.TableHeaderCell>Price Change (%)</S.TableHeaderCell>
+        </tr>
+      </S.TableHead>
+      <tbody>
+        {Object.entries(exchangesInfo)?.map(([key, value]) => (
+          <S.TableRow key={key}>
+            <td>{value.s}</td>
+            <S.TableCell>{formatNumberToFixed(value.c)}</S.TableCell>
+            <S.TableCell>{formatNumberToFixed(value.b)}</S.TableCell>
+            <S.TableCell>{formatNumberToFixed(value.a)}</S.TableCell>
+            <S.TableCell isPositive={parseFloat(value.p) * 10000}>
+              {formatPercentage(value.p)}
+            </S.TableCell>
+          </S.TableRow>
+        ))}
+      </tbody>
+    </S.Table>
+  );
+};
+
+export default ExchangeTable;
