+++ src/pages/Dashboard/components/ListSymbols/index.tsx
@@ -1,15 +1,9 @@
-import { useState, useCallback, ChangeEvent, useEffect } from "react";
+import { useState, ChangeEvent } from "react";
 import { useExchangeInfo } from "../../context/useExchangeInfo";
 import * as S from "./styles";
-import useWebSocket from "react-use-websocket";
 import { useForm, SubmitHandler } from "react-hook-form";
 import { ACTION_TYPE } from "../../context/useExchangeInfo/actions";
-
-interface WebSocketMessage {
-  stream: string;
-  data: any;
-  // TODO: increase more details about object WebSocketMessage
-}
+import ExchangeTable from "./components/ExchangeTable";
 
 type Inputs = {
   listName: string;
@@ -17,48 +11,11 @@ type Inputs = {
 };
 
 const ListSymbols = () => {
-  const [exchangesInfo, setExchangesInfo] = useState<any[]>([]);
   const [showAddNewList, setShowAddNewList] = useState(false);
 
   const { exchanges, dispatchExchanges } = useExchangeInfo();
   const [selectedOption, setSelectedOption] = useState(exchanges.currentList);
   const { register, handleSubmit } = useForm<Inputs>();
-
-  const symbolList =
-    exchanges.currentList !== "" &&
-    Object.values(exchanges.lists[exchanges.currentList])
-      ?.map((symbol: any) => symbol.symbol.toLowerCase())
-      .join("@ticker/");
-
-  const wsUrl = `wss://stream.binance.com:9443/stream?streams=${symbolList}`;
-
-  const { lastJsonMessage } = useWebSocket<WebSocketMessage>(wsUrl, {
-    onOpen: () => console.log(`Connected to App WS`),
-    onError: (event) => {
-      console.error(event);
-    },
-    onMessage: () => {
-      if (lastJsonMessage) {
-        setExchangesInfo((prev) => ({
-          ...prev,
-          [lastJsonMessage.stream]: { ...lastJsonMessage.data },
-        }));
-      }
-    },
-    shouldReconnect: () => true,
-    reconnectInterval: 10,
-  });
-
-  const formatNumberToFixed = useCallback(
-    (value: string, decimalPlaces: number = 4) => {
-      return parseFloat(value).toFixed(decimalPlaces);
-    },
-    []
-  );
-
-  const formatPercentage = useCallback((value: number) => {
-    return (value * 10000).toFixed(2);
-  }, []);
 
   const onSubmit: SubmitHandler<Inputs> = (data) => {
     dispatchExchanges({
@@ -79,10 +36,6 @@ const ListSymbols = () => {
 
     setSelectedOption(e.target.value);
   };
-
-  useEffect(() => {
-    setExchangesInfo([]);
-  }, [exchanges.currentList]);
 
   return (
     <S.Wrapper>
@@ -112,30 +65,7 @@ const ListSymbols = () => {
         )}
       </S.WrapperOfListUser>
 
-      <S.Table>
-        <S.TableHead>
-          <tr>
-            <S.TableHeaderCell>Symbol</S.TableHeaderCell>
-            <S.TableHeaderCell>Last Price</S.TableHeaderCell>
-            <S.TableHeaderCell>Bid Price</S.TableHeaderCell>
-            <S.TableHeaderCell>Ask Price</S.TableHeaderCell>
-            <S.TableHeaderCell>Price Change (%)</S.TableHeaderCell>
-          </tr>
-        </S.TableHead>
-        <tbody>
-          {Object.entries(exchangesInfo)?.map(([key, value]) => (
-            <S.TableRow key={key}>
-              <td>{value.s}</td>
-              <S.TableCell>{formatNumberToFixed(value.c)}</S.TableCell>
-              <S.TableCell>{formatNumberToFixed(value.b)}</S.TableCell>
-              <S.TableCell>{formatNumberToFixed(value.a)}</S.TableCell>
-              <S.TableCell isPositive={parseFloat(value.p) * 10000}>
-                {formatPercentage(value.p)}
-              </S.TableCell>
-            </S.TableRow>
-          ))}
-        </tbody>
-      </S.Table>
+      <ExchangeTable />
     </S.Wrapper>
   );
 };
